switch to use of Mono helper classes
severely burnt out from not understanding how to play this game, so testing is a hassle. this loads 2 files (provided to the user automatically) from [just-ero/AutoSplitHelp/mono](https://github.com/just-ero/AutoSplitHelp/tree/main/mono) to compile an assembly from which to call custom functions.

* `start` action is created according to the speedrun.com rules (with some work arounds, since no perfect events exist for those conditions).  
* `split` is obvious enough, loops over all new events if there are any, checks if the setting is enabled.
* `reset` runs when going back to the main menu.
* leaderboards don't go by in-game time (demo timing starts even before in-game time does), hence why we remove loads instead.

please approve (and/or make additions to settings) before merging; the `LiveSplit.AutoSplitters.xml` file must be edited before this can be provided to users, since the `mono.cs` and `mono_helpers.cs` files must be added to the `<URLs>`.